fix: narrow exception handlers, fix busy-spin, widen failure counter
- Narrow except Exception to except (OSError, ValueError) in both
  thread loops so programming errors propagate instead of being masked
- Replace in_waiting polling with blocking read(1) in received_loop
  to eliminate CPU busy-spin (serial timeout=1 handles the wait)
- Upgrade tcm_parse_failure_count from uint16_t to uint32_t to avoid
  silent wrap at 65535

+++ laser_engine/laser_engine.ino
@@ -182,7 +182,7 @@ uint8_t host_protocol_buf[256];
 uint16_t host_protocol_buf_length = 0;
 
 uint8_t key_status = 0;
-uint16_t tcm_parse_failure_count = 0;
+uint32_t tcm_parse_failure_count = 0;
 
 enum LEDState {
   RED,
